Refactor: 매물 등록 폼을 하나로 병합 &  분류에 따른 v-if 처리
