+++ src/modules/components/room/form/BasicInfo.vue
@@ -4,10 +4,10 @@
         <div class="row">
 
             <!-- 제목 & 주소-->
-            <NameAddress />
+            <NameAddressInput />
 
-            <!-- 임대 유형 -->
-            <div class="col-sm-6 mb-3 w-100">
+            <!-- 자취방: 임대 유형 -->
+            <div v-if="category === 'jachiroom'" class="col-sm-6 mb-3 w-100">
                 <div class="form-label pt-3 pb-2 fw-bold">임대 유형<span class="text-danger">*</span></div>
                 <div class="container row w-100">
                     <div class="form-check col-lg-3 justify-content-around">
@@ -21,18 +21,68 @@
                 </div>
             </div>
 
-            <!-- 계약금 -->
-            <Price :rentalType="rentalType" />
+            <!-- 공유주거 - 쉐어하우스: 주거 타입 -->
+            <div v-if="category === 'sharehouse'"  class="col-sm-6 mb-3 w-100">
+                <div class="form-label pt-3 pb-2 fw-bold">주거 타입<span class="text-danger">*</span></div>
+                <div class="container row w-100">
+                    <div class="form-check col-lg-3 justify-content-around">
+                        <input class="form-check-input" type="radio" id="period-one" name="ap-business-type">
+                        <label class="form-check-label" for="period-one">공유 주택</label>
+                    </div>
+                    <div class="form-check col-lg-3">
+                        <input class="form-check-input" type="radio" id="period-two" name="ap-business-type">
+                        <label class="form-check-label" for="period-two">쉐어하우스</label>
+                    </div>
+                </div>
+            </div>
 
-            <!-- 최소 계약 기간 -->
-            <label class="form-label fw-bold" for="ap-category">최소 계약 기간<span class="text-danger">*</span></label>
-            <div class="input-group mb-3">
+            <!-- 고시원 & 공유주거: 계약 최소 기간 -->
+            <div v-if="category === 'gosiwon' || category === 'sharehouse'" class="col-sm-6 mb-3 w-100">
+              <div class="form-label pt-3 pb-2 fw-bold">계약 최소 기간<span class="text-danger">*</span></div>
+              <div class="container row w-100">
+                    <div class="form-check col-lg-3 justify-content-around">
+                        <input class="form-check-input" type="radio" id="period-one" name="ap-business-type">
+                        <label class="form-check-label" for="period-one">1개월</label>
+                    </div>
+                    <div class="form-check col-lg-3">
+                        <input class="form-check-input" type="radio" id="period-two" name="ap-business-type">
+                        <label class="form-check-label" for="period-two">2개월</label>
+                    </div>
+                    <div class="form-check col-lg-3">
+                        <input class="form-check-input" type="radio" id="period-three" name="ap-business-type">
+                        <label class="form-check-label" for="period-three">3개월</label>
+                    </div>
+                    <div class="form-check col-lg-3">
+                        <input class="form-check-input" type="radio" id="period-four" name="ap-business-type">
+                        <label class="form-check-label" for="period-four">4개월 이상</label>
+                    </div>
+              </div>
+            </div>
+
+            <!-- 계약금 -->
+            <PriceInput :rentalType="rentalType" :category="category" />
+
+            <!-- 자취방: 최소 계약 기간 -->
+            <label v-if="category === 'jachiroom'" class="form-label fw-bold" for="ap-category">최소 계약 기간<span class="text-danger">*</span></label>
+            <div v-if="category === 'jachiroom'" class="input-group mb-3">
                 <input class="form-control range-slider-value-max" type="number" required>
                 <span class="input-group-text fs-base">개월</span>
             </div>
+
+            <!-- 고시원 & 공유: 이용 가능 연령 -->
+            <div v-if="category !== 'jachiroom'" class="mb-3">
+                <div class="mb-4">
+                    <label class="form-label fw-bold" for="ap-area">이용 가능 연령<span class="text-danger">*</span></label>
+                    <input class="form-control" type="number" id="ap-area" min="20" placeholder="Enter your area" value="20" required :disabled="isNoLimitAge">
+                    <div class="form-check d-flex justify-content-end pt-2 fs-sm">
+                        <input class="form-check-input" type="checkbox" id="no-age-limit" name="no-age-limit" v-model="isNoLimitAge">
+                        <label class="form-check-label px-2" for="no-age-limit">연령 제한 없음</label>
+                    </div>
+                </div>
+            </div>
           
-            <!-- 해당 층 / 전체 층 -->
-            <div class="row">
+            <!-- 자취방: 해당 층 / 전체 층 -->
+            <div v-if="category === 'jachiroom'" class="row">
                 <div class="col-sm-6 mb-3">
                   <label class="form-label fw-bold" for="ap-category">해당 층<span class="text-danger">*</span></label>
                   <input class="form-control" type="number" id="ap-area" min="-1" placeholder="매물이 위치한 층을 입력하세요" required>
@@ -43,8 +93,31 @@
                 </div>
             </div>
 
-            <!-- 방 종류 -->
-            <div class="mb-3">
+            <!-- 고시원 + 공유주거: 남녀 구분 -->
+            <div v-if="category !== 'jachiroom'" class="col-sm-6 mb-5 w-100 ">
+                <div class="form-label pt-3 pb-2 fw-bold">성별구분<span class="text-danger">*</span></div>
+                <div class="container row w-100">
+                    <div class="form-check col-lg-3 justify-content-around">
+                        <input class="form-check-input" type="radio" id="period-one" name="ap-business-type">
+                        <label class="form-check-label" for="period-one">구분 없음</label>
+                    </div>
+                    <div class="form-check col-lg-3">
+                        <input class="form-check-input" type="radio" id="period-two" name="ap-business-type">
+                        <label class="form-check-label" for="period-two">남녀 분리</label>
+                    </div>
+                    <div class="form-check col-lg-3">
+                        <input class="form-check-input" type="radio" id="period-three" name="ap-business-type">
+                        <label class="form-check-label" for="period-three">여성 전용</label>
+                    </div>
+                    <div class="form-check col-lg-3">
+                        <input class="form-check-input" type="radio" id="period-four" name="ap-business-type">
+                        <label class="form-check-label" for="period-four">남성 전용</label>
+                    </div>
+                </div>
+            </div>
+
+            <!-- 자취방: 방 종류 -->
+            <div v-if="category === 'jachiroom'" class="mb-3">
               <label class="form-label fw-bold" for="ap-category">방 종류<span class="text-danger">*</span></label>
               <select class="form-select" id="ap-category" required>
                 <option value="" disabled hidden>방 종류를 선택하세요</option>
@@ -54,8 +127,23 @@
               </select>
             </div>
 
-            <!-- 방 구조 -->
-            <div class="mb-3">
+            <!-- 쉐어하우스: 방 종류 -->
+            <div v-if="category === 'sharehouse'" class="col-sm-6 mb-5 w-100 ">
+              <div class="form-label pt-3 pb-2 fw-bold">방 개수<span class="text-danger">*</span></div>
+              <div class="mb-3 d-flex gap-1">
+                  <div class="col-lg-6">
+                      <label class="form-label" for="ap-area">1인실</label>
+                      <input class="form-control" type="number" id="ap-area" min="0" placeholder="Enter your area" value="0" required>
+                  </div>
+                  <div class="col-lg-6">
+                      <label class="form-label" for="ap-area">2인실</label>
+                      <input class="form-control" type="number" id="ap-area" min="0" placeholder="Enter your area" value="0" required>
+                  </div>
+              </div>
+            </div>
+
+            <!-- 자취방: 방 구조 -->
+            <div v-if="category === 'jachiroom'" class="mb-3">
               <label class="form-label fw-bold" for="ap-category">방 구조<span class="text-danger">*</span></label>
               <select class="form-select" id="ap-category" required>
                 <option value="" disabled hidden>방 구조를 선택하세요</option>
@@ -67,8 +155,8 @@
               </select>
             </div>
 
-            <!-- 방 수 / 욕실 수 -->
-            <div class="row">
+            <!-- 자취방: 방 수 / 욕실 수 -->
+            <div v-if="category === 'jachiroom'" class="row">
                 <div class="col-sm-6 mb-3">
                   <label class="form-label fw-bold" for="ap-category">방 개수<span class="text-danger">*</span></label>
                   <input class="form-control" type="number" id="ap-area" min="1" placeholder="방 개수를 입력하세요" required>
@@ -79,9 +167,9 @@
                 </div>
             </div>
 
-            <!-- 전용 면적 / 공급 면적 -->
-            <div class="form-label pt-3 pb-1 fw-bold">면적<span class="text-danger">*</span></div>
-            <div class="d-flex align-items-center pb-4">
+            <!-- 자취방: 전용 면적 / 공급 면적 -->
+            <div v-if="category === 'jachiroom'" class="form-label pt-3 pb-1 fw-bold">면적<span class="text-danger">*</span></div>
+            <div v-if="category === 'jachiroom'" class="d-flex align-items-center pb-4">
                 <div class="w-50 pe-2">
                     <label for="supply-area">공급 면적</label>
                     <div class="input-group">
@@ -98,8 +186,8 @@
                 </div>
             </div>
 
-            <!-- 주실 방향 -->
-            <div class="mb-3">
+            <!-- 자취방: 주실 방향 -->
+            <div v-if="category === 'jachiroom'" class="mb-3">
               <label class="form-label fw-bold" for="ap-category">주실 방향<span class="text-danger">*</span></label>
               <select class="form-select" id="ap-category" required>
                 <option value="" disabled hidden>주실 방향을 선택하세요</option>
@@ -110,15 +198,18 @@
               </select>
             </div>
 
-            <!-- 입주가능일 -->
-            <label class="form-label fw-bold">입주가능일<span class="text-danger">*</span></label>
-            <div class="input-group">
+            <!-- 자취방: 입주가능일 -->
+            <label v-if="category === 'jachiroom'" class="form-label fw-bold">입주가능일<span class="text-danger">*</span></label>
+            <div v-if="category === 'jachiroom'" class="input-group">
               <input class="form-control rounded pe-5" type="date" placeholder="입주가능일을 선택하세요" :disabled="canMoveInNow"/>
             </div>
-            <div class="form-check d-flex justify-content-end pt-2 fs-sm">
+            <div v-if="category === 'jachiroom'" class="form-check d-flex justify-content-end pt-2 fs-sm">
               <input class="form-check-input" type="checkbox" id="no-deposit" name="no-deposit-fee" v-model="canMoveInNow">
               <label class="form-check-label px-2" for="no-deposit-fee">즉시 입주 가능(협의 가능)</label>
             </div>
+
+
+
 
             <!-- 개인시설 -->
             <div class="mb-4">
@@ -224,7 +315,8 @@
                       <label class="form-check-label" for="allow-cats">외국인 가능</label>
                     </div>
                   </div>
-                  <div class="col-sm-4">
+                  <!-- 자취방: 반려동물 -->
+                  <div v-if="category === 'jachiroom'" class="col-sm-4">
                     <div class="form-check">
                       <input class="form-check-input" type="checkbox" id="allow-cats">
                       <label class="form-check-label" for="allow-cats">반려동물 가능</label>
@@ -240,16 +332,29 @@
               <textarea class="form-control" id="ap-description" rows="5" placeholder="매물에 대한 상세 정보를 입력하세요"></textarea>
             </div>
 
+            <!-- 사진 첨부 -->
+            <div class="mb-4">
+              <label class="form-label fw-bold" for="ap-pics">사진 첨부<span class="text-danger">*</span></label>
+              <input class="file-uploader file-uploader-grid mx-3" id="ap-pics" type="file" multiple data-max-files="4" required>
+              <p class="fs-sm">사진은 30MB 용량 이하에서 여러 장 첨부 가능합니다.</p>
+            </div>
         </div>
     </section>
 </template>
 
 <script setup>
 import { ref } from 'vue';
-import Price from '../Price.vue';
-import NameAddress from '../NameAddress.vue';
+import PriceInput from './PriceInput.vue';
+import NameAddressInput from './NameAddressInput.vue';
 
 const isNoLimitAge = ref(false);
 const rentalType = ref('');
 const canMoveInNow = ref(false);
+
+const props = defineProps({
+    category: {
+        type: String,
+        required: true,
+    },
+})
 </script>
