+++ src/modules/components/room/form/FacilitiesInfo.vue
