--- src/modules/components/room/form/single/SingleForm.vue
@@ -1,20 +1,0 @@
-<template>
-    <SingleBasicInfo />
-    <LoanInfo :category="category" />
-    <SingleFacilitiesInfo />
-    <BuildingInfo />
-</template>
-
-<script setup>
-import BuildingInfo from '../BuildingInfo.vue';
-import LoanInfo from '../LoanInfo.vue';
-import SingleBasicInfo from './SingleBasicInfo.vue';
-import SingleFacilitiesInfo from './SingleFacilitiesInfo.vue';
-
-const props = defineProps({
-    category: {
-        type: String,
-        required: true,
-    },
-})
-</script>
